pyproject.toml
[tool.poetry]
name = "werp_projecting_sfis"
version = "0.1.0"
description = ""
authors = ["David Robertson <david.robertson@csiro.au>", "Andrew Freebairn <andrew.freebairn@csiro.au>"]
readme = "README.md"
packages = [{include = "werp_projecting_sfis"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pytest = "^7.4.0"

pytest-github-actions-annotate-failures = "^0.2.0"
xarray = "^2024.7.0"
pymc = "^5.16.2"
netcdf4 = "^1.7.1.post1"
plotnine = "^0.13.6"
ipykernel = "^6.29.5"
numpyro = "^0.15.2"
libpython = "^0.2"
cython = "^3.0.10"
pymc-learn = "^0.0.1rc3"
arviz = "^0.19.0"
ipywidgets = "^8.1.3"
matplotlib = "3.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
